Reorganize quantizer definitions
